refactor: 엔티티 위치 수정
