+++ src/main/java/com/example/todo/domain/repository/TaskApiRepository.java
@@ -1,6 +1,7 @@
 package com.example.todo.domain.repository;
 
-import com.example.todo.api.TaskApiEntity;
+import com.example.todo.domain.entity.task.TaskApiEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
-    public interface TaskApiRepository extends JpaRepository<TaskApiEntity, Long> {
-    }
+
+public interface TaskApiRepository extends JpaRepository<TaskApiEntity, Long> {
+}
