+++ src/main/java/com/example/todo/dto/TaskApiDto.java
@@ -1,6 +1,6 @@
 package com.example.todo.dto;
 
-import com.example.todo.api.TaskApiEntity;
+import com.example.todo.domain.entity.task.TaskApiEntity;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
